TimedMemberStateFactory refactored
