+++ hazelcast/src/main/java/com/hazelcast/instance/DefaultNodeExtension.java
@@ -20,6 +20,8 @@ import com.hazelcast.config.Config;
 import com.hazelcast.config.SerializationConfig;
 import com.hazelcast.core.PartitioningStrategy;
 import com.hazelcast.logging.ILogger;
+import com.hazelcast.management.DefaultTimedMemberStateFactory;
+import com.hazelcast.management.TimedMemberStateFactory;
 import com.hazelcast.nio.ClassLoaderUtil;
 import com.hazelcast.nio.IOService;
 import com.hazelcast.nio.MemberSocketInterceptor;
@@ -118,8 +120,13 @@ public class DefaultNodeExtension implements NodeExtension {
     }
 
     @Override
-    public WanReplicationService geWanReplicationService() {
+    public WanReplicationService getWanReplicationService() {
         return new WanReplicationServiceImpl(node);
+    }
+
+    @Override
+    public TimedMemberStateFactory getTimedMemberStateFactory() {
+        return new DefaultTimedMemberStateFactory(node.hazelcastInstance);
     }
 
     @Override
