+++ hazelcast/src/main/java/com/hazelcast/monitor/impl/LocalMapStatsImpl.java
@@ -18,14 +18,9 @@ package com.hazelcast.monitor.impl;
 
 import com.eclipsesource.json.JsonObject;
 import com.eclipsesource.json.JsonValue;
-import com.hazelcast.map.impl.MapDataSerializerHook;
 import com.hazelcast.monitor.LocalMapStats;
-import com.hazelcast.nio.ObjectDataInput;
-import com.hazelcast.nio.ObjectDataOutput;
-import com.hazelcast.nio.serialization.IdentifiedDataSerializable;
 import com.hazelcast.util.Clock;
 
-import java.io.IOException;
 import java.util.concurrent.atomic.AtomicLongFieldUpdater;
 
 import static com.hazelcast.util.JsonUtil.getInt;
@@ -34,8 +29,7 @@ import static com.hazelcast.util.JsonUtil.getLong;
 /**
  * Default implementation of {@link LocalMapStats}
  */
-public class LocalMapStatsImpl
-        implements LocalMapStats, IdentifiedDataSerializable {
+public class LocalMapStatsImpl implements LocalMapStats {
 
     private static final AtomicLongFieldUpdater<LocalMapStatsImpl> LAST_ACCESS_TIME_UPDATER = AtomicLongFieldUpdater
             .newUpdater(LocalMapStatsImpl.class, "lastAccessTime");
@@ -121,72 +115,6 @@ public class LocalMapStatsImpl
         dirtyEntryCount = 0;
         backupCount = 0;
         hits = 0;
-    }
-
-    @Override
-    public void writeData(ObjectDataOutput out)
-            throws IOException {
-        out.writeLong(getCount);
-        out.writeLong(putCount);
-        out.writeLong(removeCount);
-        out.writeLong(numberOfOtherOperations);
-        out.writeLong(numberOfEvents);
-        out.writeLong(lastAccessTime);
-        out.writeLong(lastUpdateTime);
-        out.writeLong(hits);
-        out.writeLong(ownedEntryCount);
-        out.writeLong(backupEntryCount);
-        out.writeInt(backupCount);
-        out.writeLong(ownedEntryMemoryCost);
-        out.writeLong(backupEntryMemoryCost);
-        out.writeLong(creationTime);
-        out.writeLong(lockedEntryCount);
-        out.writeLong(dirtyEntryCount);
-        out.writeLong(totalGetLatencies);
-        out.writeLong(totalPutLatencies);
-        out.writeLong(totalRemoveLatencies);
-        out.writeLong(maxGetLatency);
-        out.writeLong(maxPutLatency);
-        out.writeLong(maxRemoveLatency);
-        out.writeLong(heapCost);
-        boolean hasNearCache = nearCacheStats != null;
-        out.writeBoolean(hasNearCache);
-        if (hasNearCache) {
-            nearCacheStats.writeData(out);
-        }
-    }
-
-    @Override
-    public void readData(ObjectDataInput in)
-            throws IOException {
-        GET_COUNT_UPDATER.set(this, in.readLong());
-        PUT_COUNT_UPDATER.set(this, in.readLong());
-        REMOVE_COUNT_UPDATER.set(this, in.readLong());
-        NUMBER_OF_OTHER_OPERATIONS_UPDATER.set(this, in.readLong());
-        NUMBER_OF_EVENTS_UPDATER.set(this, in.readLong());
-        LAST_ACCESS_TIME_UPDATER.set(this, in.readLong());
-        LAST_UPDATE_TIME_UPDATER.set(this, in.readLong());
-        HITS_UPDATER.set(this, in.readLong());
-        ownedEntryCount = in.readLong();
-        backupEntryCount = in.readLong();
-        backupCount = in.readInt();
-        ownedEntryMemoryCost = in.readLong();
-        backupEntryMemoryCost = in.readLong();
-        creationTime = in.readLong();
-        lockedEntryCount = in.readLong();
-        dirtyEntryCount = in.readLong();
-        TOTAL_GET_LATENCIES_UPDATER.set(this, in.readLong());
-        TOTAL_PUT_LATENCIES_UPDATER.set(this, in.readLong());
-        TOTAL_REMOVE_LATENCIES_UPDATER.set(this, in.readLong());
-        MAX_GET_LATENCY_UPDATER.set(this, in.readLong());
-        MAX_PUT_LATENCY_UPDATER.set(this, in.readLong());
-        MAX_REMOVE_LATENCY_UPDATER.set(this, in.readLong());
-        heapCost = in.readLong();
-        boolean hasNearCache = in.readBoolean();
-        if (hasNearCache) {
-            nearCacheStats = new NearCacheStatsImpl();
-            nearCacheStats.readData(in);
-        }
     }
 
     @Override
@@ -402,16 +330,6 @@ public class LocalMapStatsImpl
 
     public void setNearCacheStats(NearCacheStatsImpl nearCacheStats) {
         this.nearCacheStats = nearCacheStats;
-    }
-
-    @Override
-    public int getFactoryId() {
-        return MapDataSerializerHook.F_ID;
-    }
-
-    @Override
-    public int getId() {
-        return MapDataSerializerHook.MAP_STATS;
     }
 
     public JsonObject toJson() {
