+++ hazelcast/src/main/java/com/hazelcast/monitor/impl/LocalReplicatedMapStatsImpl.java
@@ -18,12 +18,8 @@ package com.hazelcast.monitor.impl;
 
 import com.eclipsesource.json.JsonObject;
 import com.hazelcast.monitor.LocalReplicatedMapStats;
-import com.hazelcast.nio.ObjectDataInput;
-import com.hazelcast.nio.ObjectDataOutput;
-import com.hazelcast.nio.serialization.IdentifiedDataSerializable;
-import com.hazelcast.replicatedmap.impl.operation.ReplicatedMapDataSerializerHook;
 import com.hazelcast.util.Clock;
-import java.io.IOException;
+
 import java.util.concurrent.atomic.AtomicLongFieldUpdater;
 
 import static com.hazelcast.util.JsonUtil.getLong;
@@ -33,7 +29,7 @@ import static com.hazelcast.util.JsonUtil.getLong;
  * able to transform those between wire format and instance view
  */
 public class LocalReplicatedMapStatsImpl
-        implements LocalReplicatedMapStats, IdentifiedDataSerializable {
+        implements LocalReplicatedMapStats {
 
     //CHECKSTYLE:OFF
     private static final AtomicLongFieldUpdater<LocalReplicatedMapStatsImpl> LAST_ACCESS_TIME_UPDATER = AtomicLongFieldUpdater
@@ -90,50 +86,6 @@ public class LocalReplicatedMapStatsImpl
 
     public LocalReplicatedMapStatsImpl() {
         creationTime = Clock.currentTimeMillis();
-    }
-
-    @Override
-    public void writeData(ObjectDataOutput out)
-            throws IOException {
-        out.writeLong(getCount);
-        out.writeLong(putCount);
-        out.writeLong(removeCount);
-        out.writeLong(numberOfOtherOperations);
-        out.writeLong(numberOfEvents);
-        out.writeLong(numberOfReplicationEvents);
-        out.writeLong(lastAccessTime);
-        out.writeLong(lastUpdateTime);
-        out.writeLong(hits);
-        out.writeLong(ownedEntryCount);
-        out.writeLong(creationTime);
-        out.writeLong(totalGetLatencies);
-        out.writeLong(totalPutLatencies);
-        out.writeLong(totalRemoveLatencies);
-        out.writeLong(maxGetLatency);
-        out.writeLong(maxPutLatency);
-        out.writeLong(maxRemoveLatency);
-    }
-
-    @Override
-    public void readData(ObjectDataInput in)
-            throws IOException {
-        GET_COUNT_UPDATER.set(this, in.readLong());
-        PUT_COUNT_UPDATER.set(this, in.readLong());
-        REMOVE_COUNT_UPDATER.set(this, in.readLong());
-        NUMBER_OF_OTHER_OPERATIONS_UPDATER.set(this, in.readLong());
-        NUMBER_OF_EVENTS_UPDATER.set(this, in.readLong());
-        NUMBER_OF_REPLICATION_EVENTS_UPDATER.set(this, in.readLong());
-        LAST_ACCESS_TIME_UPDATER.set(this, in.readLong());
-        LAST_UPDATE_TIME_UPDATER.set(this, in.readLong());
-        HITS_UPDATER.set(this, in.readLong());
-        ownedEntryCount = in.readLong();
-        creationTime = in.readLong();
-        TOTAL_GET_LATENCIES_UPDATER.set(this, in.readLong());
-        TOTAL_PUT_LATENCIES_UPDATER.set(this, in.readLong());
-        TOTAL_REMOVE_LATENCIES_UPDATER.set(this, in.readLong());
-        MAX_GET_LATENCY_UPDATER.set(this, in.readLong());
-        MAX_PUT_LATENCY_UPDATER.set(this, in.readLong());
-        MAX_REMOVE_LATENCY_UPDATER.set(this, in.readLong());
     }
 
     @Override
@@ -392,13 +344,4 @@ public class LocalReplicatedMapStatsImpl
                 + ", creationTime=" + creationTime + '}';
     }
 
-    @Override
-    public int getFactoryId() {
-        return ReplicatedMapDataSerializerHook.F_ID;
-    }
-
-    @Override
-    public int getId() {
-        return ReplicatedMapDataSerializerHook.MAP_STATS;
-    }
 }
