+++ hazelcast/src/main/java/com/hazelcast/nio/serialization/SerializationService.java
@@ -18,7 +18,6 @@ package com.hazelcast.nio.serialization;
 
 import com.hazelcast.core.ManagedContext;
 import com.hazelcast.core.PartitioningStrategy;
-import com.hazelcast.memory.MemoryManager;
 import com.hazelcast.nio.BufferObjectDataInput;
 import com.hazelcast.nio.BufferObjectDataOutput;
 import com.hazelcast.nio.ObjectDataInput;
@@ -76,6 +75,4 @@ public interface SerializationService {
     void push(BufferObjectDataOutput out);
 
     void destroy();
-
-    MemoryManager getMemoryManager();
 }
